Allow users to set env vars (#83)
## Why?

A Nix module should be able to set specific environment variables.

## Changes

Allow the module author to specify env vars as a map. These env vars is
in addition to the `PATH` variable automatically getting set based on
the bin directories of the specified packages field.

## Testing

1. `cd pkgs/moduleit`
2. `./moduleit.sh example.nix`
3. `cat ./result | jq` and see that env has both the `PATH` variable
correctly set and the `FOO` variable set.