Set default  (#85)
## Why

I set the default channel in default.nix to 22.11, but that broke the
build.sh script and caused all channels to be 22.11. This broke Pygame
in Canary.

## Change

Updated the default channel in the build script.

## Testing

Deploy cacache disk again, and see that each channel correctly
references its own code. See that Pygame is fixed.