Add dart 2.10 (#102)
The legacy polygott images had Dart 2.9 so this is technically an
update, but 2.10 only adds features so it's no big deal to upgrade
polygott users to 2.10 in nix migrations

HOWEVER, dart 2.12 had a major breaking change - so big of a change, in
fact, that [dart 2.11 was
skipped](https://groups.google.com/a/dartlang.org/g/misc/c/GMHD6LOLJ1I?pli=1).
Why would Google do this? Why are we in this world? What even is
versioning? How come we can't have nice things? Must we go on like this?
Must all of our software break for no reason at the whims of Google and
their anti-versioning gods? What is the meaning of versions? What is the
meaning of life?