Experiment with importing modules (#131)
# Why?

Experimented with using Nix modules to compose modules.

`typescript-language-server.nix` is a module, and it was imported into
nodejs.nix, web.nix and bun.nix. The [module
system](https://github.com/replit/nixpkgs-replit/blob/main/pkgs/moduleit/entrypoint.nix#L5)
performs a deep merge of the imported modules into the main module,
which may or may not be desirable. Data types that are mergeable are
lists and attributes sets (like JS objects). Strings, booleans and
numbers are not mergable and will conflict if they exist in both the
imported and the importing modules. For that reason, I have left out
name and version in `typescript-language-server.nix`

## Changes

1. Added importable module `typescript-language-server.nix`
2. nodejs, web, and bun modules make use of the new module, overriding
the extensions field when needed
3. Changed the language field for bun's lsp and runner to `javascript`
because I think that's more apt. While for upm, we leave the language as
`bun`

## Testing

```
nix-build -A replitPackages.modules.nodejs # or the other 2
cat result | jq
```
You'll see the typescript server and it's extensions list as expected.