Runner priority (#137)
# Why?

Add runner priority field to allow a module to prefer some runners vs
others. This came up in the C and C++ modules where multiple runners can
be applicable at the same time, but there isn't a way to specify one vs
the other and the result came down to chance. And for users that don't
have `flag-toolchain-multiple-runners`, they won't have a way to switch
the runner if we got it wrong.

## Changes

Added a priority field to runners. Lower means higher. The plan is to
have `.replit/interp` at 0 and `.replit/run` at 1, and then any priority
set in a Nix module should be 1 or greater. If priority is not set, it
means infinity.