Move data directory out of REPL_HOME (#156)
# Why

To solve [issues with Java
LSP](https://linear.app/replit/issue/DX-377/java-files-not-being-recognized-as-project-files-causing-issues-with)
where java files were labeled with warning that say they are non-project
files and therefore you'd only get syntax error assistance.

I found that there's an error when JDT starts and tries to import Maven
projects:

```
Failed to import projects
Invalid project description.
org.eclipse.core.internal.resources.ResourceException: Invalid project description.
	at org.eclipse.core.internal.resources.Project.checkDescription(Project.java:211)
.
.
.
Contains: OK
Contains: /home/runner/PlayfulGhostwhiteCallbacks overlaps the workspace location: /home/runner/PlayfulGhostwhiteCallbacks/.cache/jdt-language-server/data/home/runner/PlayfulGhostwhiteCallbacks/
```

because the data directory we've specified for JDT is within the
REPL_HOME directory, which is used as the Maven project directory, and
this is not allowed. A search in the source code didn't give me a good
way to allow this as the logic is deep in an Eclipse library and not the
JDT source code. So unfortunately for now I propose moving the data
directory to $HOME/.cache, but not the config directory, which contains
18M worth of jars.

# Changes

Split runtimePath into configDirPath and dataDirPath. Set dataDirPath to
$HOME/.cache unconditionally where as configDirPath still can be
controlled by $XDG_CACHE_HOME. So for a normal Java repl we'd expect
config dir to be $REPL_HOME/.cache/jdt-language-server/config where as
data dir should be $HOME/.cache/jdt-language-server/data.

In order to get rid of all warnings of source files not being project
files, we'd also need to change the Java template to move Main.java into
the `src/main/java` subdirectory as per Maven conventions. However, that
breaks the debugger. For that, that there's another fix.

# Test

When running jdt-language-server for the first time, it should create
the data and config directory in the corresponding places. You can
remove those directories if they are already there to test.